refactor(core): replace corehr with platform-admin module and core-ui layer
Made-with: Cursor
