+++ Frontend/apps/core/public/stitch/organizations-list.png
