+++ Frontend/apps/core/public/stitch/public-invite.png
