+++ Frontend/apps/core/src/app/(core)/design-system/page.tsx
@@ -1,0 +1,5 @@
+import { DesignSystemOverview } from "@/modules/platform-admin/components/design-system-overview";
+
+export default function DesignSystemPage() {
+  return <DesignSystemOverview />;
+}
