+++ Frontend/apps/core/src/app/(core)/organizations/new/page.tsx
@@ -1,0 +1,5 @@
+import { CreateOrganizationForm } from "@/modules/platform-admin/components/create-organization-form";
+
+export default function CreateOrganizationPage() {
+  return <CreateOrganizationForm />;
+}
