+++ Frontend/apps/core/src/app/(core)/organizations/page.tsx
@@ -1,0 +1,5 @@
+import { OrganizationsListView } from "@/modules/platform-admin/components/organizations-list-view";
+
+export default function OrganizationsListPage() {
+  return <OrganizationsListView />;
+}
