--- Frontend/apps/core/src/app/corehr/design-system/page.tsx
@@ -1,5 +1,0 @@
-import { DesignSystemOverview } from "@/modules/corehr/components/design-system-overview";
-
-export default function DesignSystemPage() {
-  return <DesignSystemOverview />;
-}
