--- Frontend/apps/core/src/app/corehr/organizations/new/page.tsx
@@ -1,5 +1,0 @@
-import { CreateOrganizationForm } from "@/modules/corehr/components/create-organization-form";
-
-export default function CreateOrganizationPage() {
-  return <CreateOrganizationForm />;
-}
