--- Frontend/apps/core/src/app/corehr/organizations/page.tsx
@@ -1,5 +1,0 @@
-import { OrganizationsListView } from "@/modules/corehr/components/organizations-list-view";
-
-export default function OrganizationsListPage() {
-  return <OrganizationsListView />;
-}
