--- Frontend/apps/core/src/app/corehr/page.tsx
@@ -1,5 +1,0 @@
-import { redirect } from "next/navigation";
-
-export default function CoreHrIndexPage() {
-  redirect("/corehr/organizations");
-}
