+++ Frontend/apps/core/src/components/core-ui/index.ts
@@ -1,0 +1,13 @@
+/**
+ * Fusion Core app — composes `@repo/ui` primitives with shared field + CTA tokens (`core-ui-classes`).
+ */
+export { CoreInput } from "./core-input";
+export type { CoreInputProps } from "./core-input";
+export { CoreTextarea } from "./core-textarea";
+export type { CoreTextareaProps } from "./core-textarea";
+export { CorePrimaryButton } from "./core-primary-button";
+export {
+  coreFieldClassName,
+  coreFieldReadOnlyClassName,
+  corePrimaryButtonClassName,
+} from "@/lib/core-ui-classes";
