--- Frontend/apps/core/src/modules/corehr/data/mock-organizations.ts
@@ -1,105 +1,0 @@
-import type { Organization } from "../types/organization";
-
-export const MOCK_ORGANIZATIONS: Organization[] = [
-  {
-    id: "org-acme-monolithics",
-    name: "Acme Monolithics",
-    initials: "AM",
-    lifecycle: "active",
-    adminStatus: "Verified",
-    userCount: 1420,
-    pendingInvites: 12,
-    lastActivity: "2 mins ago",
-    createdAt: "Jun 12, 2023",
-    description:
-      "Enterprise logistics management system. Operations at full capacity across regions.",
-    primaryAdminName: "Sarah Chen",
-    primaryAdminEmail: "s.chen@acmemonolithics.com",
-    inviteSentAt: "Jun 12, 2023, 08:00 AM",
-    inviteExpiresAt: "Jun 19, 2023, 08:00 AM",
-    onboardingProgressPercent: 100,
-    onboardingStageTitle: "Production",
-    onboardingStageSubtitle: "Live",
-  },
-  {
-    id: "org-global-cyberdyne",
-    name: "Global Cyberdyne",
-    initials: "GC",
-    lifecycle: "attention",
-    adminStatus: "Unconfirmed",
-    userCount: 0,
-    pendingInvites: 1,
-    lastActivity: null,
-    createdAt: "Jan 12, 2024",
-    description:
-      "Security automation pilot. Awaiting confirmation after invite issue.",
-    primaryAdminName: "Johnathan Doe",
-    primaryAdminEmail: "j.doe@globalcyberdyne.com",
-    inviteSentAt: "Jan 12, 2024, 09:12 AM",
-    inviteExpiresAt: "Jan 19, 2024, 09:12 AM",
-    onboardingProgressPercent: 25,
-    onboardingStageTitle: "Attention",
-    onboardingStageSubtitle: "Next: Resend invite",
-  },
-  {
-    id: "org-vandelay",
-    name: "Vandelay Systems",
-    initials: "VS",
-    lifecycle: "invited",
-    adminStatus: "Awaiting Login",
-    userCount: 0,
-    pendingInvites: 5,
-    lastActivity: "4 days ago",
-    createdAt: "Oct 24, 2023",
-    description:
-      "Import/export analytics. Currently in the initial onboarding phase awaiting administrative confirmation.",
-    primaryAdminName: "Johnathan Doe",
-    primaryAdminEmail: "j.doe@vandelaysystems.com",
-    inviteSentAt: "Oct 24, 09:12 AM",
-    inviteExpiresAt: "Oct 31, 09:12 AM",
-    onboardingProgressPercent: 25,
-    onboardingStageTitle: "Awaiting Acceptance",
-    onboardingStageSubtitle: "Next: Initial Login",
-  },
-  {
-    id: "org-stark",
-    name: "Stark Tectonics",
-    initials: "ST",
-    lifecycle: "suspended",
-    adminStatus: "Revoked",
-    userCount: 84,
-    pendingInvites: 0,
-    lastActivity: "3 months ago",
-    createdAt: "May 10, 2022",
-    description: "Legacy tenant suspended pending audit.",
-    primaryAdminName: "Tony Stark",
-    primaryAdminEmail: "t.stark@starktectonics.com",
-    inviteSentAt: "May 10, 2022, 10:00 AM",
-    inviteExpiresAt: "May 17, 2022, 10:00 AM",
-    onboardingProgressPercent: 60,
-    onboardingStageTitle: "Suspended",
-    onboardingStageSubtitle: "Governance hold",
-  },
-];
-
-export const MOCK_ORG_STATS = {
-  totalAssets: 1284,
-  attentionNeeded: 12,
-  invitedPending: 48,
-  activeUsersLabel: "8.4k",
-} as const;
-
-export const MOCK_OPERATIONAL_LOGS = [
-  {
-    id: "log-1",
-    tone: "primary" as const,
-    body: "System Admin initiated a massive import for Wayne Enterprises.",
-    meta: "14:22 GMT · Batch #9283",
-  },
-  {
-    id: "log-2",
-    tone: "error" as const,
-    body: "Global Cyberdyne first admin invite expired. Automatic alert triggered.",
-    meta: "09:15 GMT · System Hook",
-  },
-] as const;
