--- Frontend/apps/core/src/modules/corehr/index.ts
@@ -1,10 +1,0 @@
-/**
- * CoreHR feature module — self-contained under `modules/corehr` for future
- * extraction into a separate package or micro-frontend.
- */
-export type { Organization, OrganizationLifecycle } from "./types/organization";
-export {
-  OrganizationsProvider,
-  useOrganizations,
-} from "./context/organizations-context";
-export { MOCK_ORGANIZATIONS } from "./data/mock-organizations";
