+++ Frontend/apps/core/src/modules/platform-admin/components/lifecycle-badge.tsx
