+++ Frontend/apps/core/src/modules/platform-admin/data/mock-organizations.ts
@@ -1,0 +1,253 @@
+import type { Organization } from "../types/organization";
+
+export const MOCK_ORGANIZATIONS: Organization[] = [
+  {
+    id: "org-acme-monolithics",
+    name: "Acme Monolithics",
+    initials: "AM",
+    lifecycle: "active",
+    adminStatus: "Verified",
+    userCount: 1420,
+    pendingInvites: 12,
+    lastActivity: "2 mins ago",
+    createdAt: "Jun 12, 2023",
+    description:
+      "Enterprise logistics management system. Operations at full capacity across regions.",
+    primaryAdminName: "Sarah Chen",
+    primaryAdminEmail: "s.chen@acmemonolithics.com",
+    inviteSentAt: "Jun 12, 2023, 08:00 AM",
+    inviteExpiresAt: "Jun 19, 2023, 08:00 AM",
+    onboardingProgressPercent: 100,
+    onboardingStageTitle: "Production",
+    onboardingStageSubtitle: "Live",
+  },
+  {
+    id: "org-global-cyberdyne",
+    name: "Global Cyberdyne",
+    initials: "GC",
+    lifecycle: "attention",
+    adminStatus: "Unconfirmed",
+    userCount: 0,
+    pendingInvites: 1,
+    lastActivity: null,
+    createdAt: "Jan 12, 2024",
+    description:
+      "Security automation pilot. Awaiting confirmation after invite issue.",
+    primaryAdminName: "Johnathan Doe",
+    primaryAdminEmail: "j.doe@globalcyberdyne.com",
+    inviteSentAt: "Jan 12, 2024, 09:12 AM",
+    inviteExpiresAt: "Jan 19, 2024, 09:12 AM",
+    onboardingProgressPercent: 25,
+    onboardingStageTitle: "Attention",
+    onboardingStageSubtitle: "Next: Resend invite",
+  },
+  {
+    id: "org-vandelay",
+    name: "Vandelay Systems",
+    initials: "VS",
+    lifecycle: "invited",
+    adminStatus: "Awaiting Login",
+    userCount: 0,
+    pendingInvites: 5,
+    lastActivity: "4 days ago",
+    createdAt: "Oct 24, 2023",
+    description:
+      "Import/export analytics. Currently in the initial onboarding phase awaiting administrative confirmation.",
+    primaryAdminName: "Johnathan Doe",
+    primaryAdminEmail: "j.doe@vandelaysystems.com",
+    inviteSentAt: "Oct 24, 09:12 AM",
+    inviteExpiresAt: "Oct 31, 09:12 AM",
+    onboardingProgressPercent: 25,
+    onboardingStageTitle: "Awaiting Acceptance",
+    onboardingStageSubtitle: "Next: Initial Login",
+  },
+  {
+    id: "org-stark",
+    name: "Stark Tectonics",
+    initials: "ST",
+    lifecycle: "suspended",
+    adminStatus: "Revoked",
+    userCount: 84,
+    pendingInvites: 0,
+    lastActivity: "3 months ago",
+    createdAt: "May 10, 2022",
+    description: "Legacy tenant suspended pending audit.",
+    primaryAdminName: "Tony Stark",
+    primaryAdminEmail: "t.stark@starktectonics.com",
+    inviteSentAt: "May 10, 2022, 10:00 AM",
+    inviteExpiresAt: "May 17, 2022, 10:00 AM",
+    onboardingProgressPercent: 60,
+    onboardingStageTitle: "Suspended",
+    onboardingStageSubtitle: "Governance hold",
+  },
+  {
+    id: "org-wayne-enterprises",
+    name: "Wayne Enterprises",
+    initials: "WE",
+    lifecycle: "active",
+    adminStatus: "Verified",
+    userCount: 890,
+    pendingInvites: 2,
+    lastActivity: "1 hour ago",
+    createdAt: "Mar 2, 2024",
+    description: "Conglomerate operations.",
+    primaryAdminName: "Lucius Fox",
+    primaryAdminEmail: "l.fox@wayneenterprises.com",
+    inviteSentAt: "Mar 2, 2024",
+    inviteExpiresAt: "Mar 9, 2024",
+    onboardingProgressPercent: 100,
+    onboardingStageTitle: "Production",
+    onboardingStageSubtitle: "Live",
+  },
+  {
+    id: "org-initech",
+    name: "Initech",
+    initials: "IN",
+    lifecycle: "invited",
+    adminStatus: "Awaiting Login",
+    userCount: 0,
+    pendingInvites: 1,
+    lastActivity: null,
+    createdAt: "Feb 1, 2025",
+    description: "Awaiting first admin.",
+    primaryAdminName: "Peter Gibbons",
+    primaryAdminEmail: "p.gibbons@initech.com",
+    inviteSentAt: "Feb 1, 2025",
+    inviteExpiresAt: "Feb 8, 2025",
+    onboardingProgressPercent: 15,
+    onboardingStageTitle: "Provisioning",
+    onboardingStageSubtitle: "Invite sent",
+  },
+  {
+    id: "org-paper-street",
+    name: "Paper Street Soap Co.",
+    initials: "PS",
+    lifecycle: "attention",
+    adminStatus: "Unconfirmed",
+    userCount: 0,
+    pendingInvites: 1,
+    lastActivity: null,
+    createdAt: "Jan 15, 2025",
+    description: "Invite expired — needs resend.",
+    primaryAdminName: "Tyler Durden",
+    primaryAdminEmail: "tyler@paperstreet.com",
+    inviteSentAt: "Jan 15, 2025",
+    inviteExpiresAt: "Jan 22, 2025",
+    onboardingProgressPercent: 10,
+    onboardingStageTitle: "Attention",
+    onboardingStageSubtitle: "Expired invite",
+  },
+  {
+    id: "org-dunder-mifflin",
+    name: "Dunder Mifflin",
+    initials: "DM",
+    lifecycle: "active",
+    adminStatus: "Verified",
+    userCount: 412,
+    pendingInvites: 0,
+    lastActivity: "Yesterday",
+    createdAt: "Nov 8, 2022",
+    description: "Regional paper distribution.",
+    primaryAdminName: "Michael Scott",
+    primaryAdminEmail: "m.scott@dundermifflin.com",
+    onboardingProgressPercent: 100,
+    onboardingStageTitle: "Production",
+    onboardingStageSubtitle: "Live",
+  },
+  {
+    id: "org-hooli",
+    name: "Hooli",
+    initials: "HO",
+    lifecycle: "active",
+    adminStatus: "Verified",
+    userCount: 2100,
+    pendingInvites: 4,
+    lastActivity: "30 mins ago",
+    createdAt: "Aug 1, 2021",
+    description: "Search and cloud.",
+    primaryAdminName: "Gavin Belson",
+    primaryAdminEmail: "g.belson@hooli.com",
+    onboardingProgressPercent: 100,
+    onboardingStageTitle: "Production",
+    onboardingStageSubtitle: "Live",
+  },
+  {
+    id: "org-pied-piper",
+    name: "Pied Piper",
+    initials: "PP",
+    lifecycle: "invited",
+    adminStatus: "Awaiting Login",
+    userCount: 0,
+    pendingInvites: 3,
+    lastActivity: null,
+    createdAt: "Dec 12, 2024",
+    description: "Middle-out compression.",
+    primaryAdminName: "Richard Hendricks",
+    primaryAdminEmail: "rhendricks@piedpiper.com",
+    inviteSentAt: "Dec 12, 2024",
+    inviteExpiresAt: "Dec 19, 2024",
+    onboardingProgressPercent: 40,
+    onboardingStageTitle: "Onboarding",
+    onboardingStageSubtitle: "Awaiting acceptance",
+  },
+  {
+    id: "org-bluth-co",
+    name: "Bluth Company",
+    initials: "BC",
+    lifecycle: "suspended",
+    adminStatus: "Suspended",
+    userCount: 12,
+    pendingInvites: 0,
+    lastActivity: "1 year ago",
+    createdAt: "Apr 4, 2003",
+    description: "Frozen banana stand (legacy).",
+    primaryAdminName: "Michael Bluth",
+    primaryAdminEmail: "m.bluth@bluth.com",
+    inviteSentAt: "Apr 4, 2003",
+    inviteExpiresAt: "Apr 11, 2003",
+    onboardingProgressPercent: 20,
+    onboardingStageTitle: "Suspended",
+    onboardingStageSubtitle: "Audit",
+  },
+  {
+    id: "org-wonka",
+    name: "Wonka Industries",
+    initials: "WI",
+    lifecycle: "active",
+    adminStatus: "Verified",
+    userCount: 340,
+    pendingInvites: 1,
+    lastActivity: "5 mins ago",
+    createdAt: "Jul 20, 2023",
+    description: "Confectionery R&D.",
+    primaryAdminName: "Willy Wonka",
+    primaryAdminEmail: "w.wonka@wonka.com",
+    inviteSentAt: "Jul 20, 2023",
+    inviteExpiresAt: "Jul 27, 2023",
+    onboardingProgressPercent: 100,
+    onboardingStageTitle: "Production",
+    onboardingStageSubtitle: "Live",
+  },
+];
+
+export const MOCK_ORG_STATS = {
+  totalAssets: 1284,
+  attentionNeeded: 12,
+  invitedPending: 48,
+  activeUsersLabel: "8.4k",
+} as const;
+
+export const MOCK_OPERATIONAL_LOGS = [
+  {
+    id: "log-1",
+    tone: "primary" as const,
+    body: "System Admin initiated a massive import for Wayne Enterprises.",
+    meta: "14:22 GMT · Batch #9283",
+  },
+  {
+    id: "log-2",
+    tone: "error" as const,
+    body: "Global Cyberdyne first admin invite expired. Automatic alert triggered.",
+    meta: "09:15 GMT · System Hook",
+  },
+] as const;
