+++ Frontend/packages/ui/package.json
@@ -16,6 +16,7 @@
     "type-check": "tsc --noEmit"
   },
   "devDependencies": {
+    "next": "^15.1.0",
     "@repo/eslint-config": "workspace:*",
     "@repo/tsconfig": "workspace:*",
     "@types/react": "^19.0.0",
@@ -44,6 +45,7 @@
     "tailwind-merge": "^2.6.0"
   },
   "peerDependencies": {
+    "next": "^15.0.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
   }
